# public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Fleet Finance Manger",
        "description": "Fleet Finance Manger - API for Android and IOS Development",
        "termsOfService": "All rights reserved to Greefi Technologies",
        "contact": {
            "email": "hello@greefitech.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:9000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/expenseType/create": {
            "post": {
                "tags": [
                    "ExpenseType"
                ],
                "description": "Create New Expense Type",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "expenseType",
                        "in": "query",
                        "description": "Enter Expense Type",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Invalid Data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/expenseTypes": {
            "get": {
                "tags": [
                    "ExpenseType"
                ],
                "description": "Return all Expense Type",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Invalid Data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/expenseType/{id}/edit": {
            "get": {
                "tags": [
                    "ExpenseType"
                ],
                "description": "Return particular Expense Type",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Invalid Data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/expenseType/{id}/update": {
            "post": {
                "tags": [
                    "ExpenseType"
                ],
                "description": "Create New Expense Type",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "expenseType",
                        "in": "query",
                        "description": "Enter Expense Type",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Invalid Data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/expenseType/{id}/delete": {
            "delete": {
                "tags": [
                    "ExpenseType"
                ],
                "description": "Delete Expense Type Data",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Invalid Data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
}